backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
firebase-admin==6.5.0
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-cloud-aiplatform==1.72.0
google-generativeai==0.8.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
mediapipe==0.10.18
numpy==1.26.4
Pillow==10.4.0
beautifulsoup4==4.12.3
httpx==0.27.2
qrcode==7.4.2
python-jose==3.3.0
